DLS: Rapatriement de l'interpretation entrées/sorties analogiques au niveau du thread IO
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@5898 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/db.c
@@ -364,7 +364,7 @@
              "Libere_DB_SQL: Deconnexion effective (DB%07d), Nbr_requete_en_cours=%d", db->id, taille );
     g_free( db );
     *adr_db = NULL;
-    Dls_data_set_AI ( "SYS", "DB_NBR_REQUEST", (gpointer)&ai_nbr_dbrequest, taille, TRUE );
+    Dls_data_set_AI ( "SYS", "DB_NBR_REQUEST", (gpointer)&ai_nbr_dbrequest, (gdouble)taille, TRUE );
   }
 /******************************************************************************************************************************/
 /* Lancer_requete_SQL : lance une requete en parametre, sur la structure de reférence                                         */
